Adds support FancyTree to display Popovers and other improvements
Fixed the Fancytree code so it renders the PopOver components instead of
the JavaScript implementation. Also fixed the code so we can add
multiple FancyTree components to the flyout.

Adds open to current node on tree navigation

If the current domain and current Feature ID matches the tree sets the
feature ID to the current Feature ID so it can be opened and set as
active.